requirements.txt
transformers
faiss-cpu
sentence-transformers
flask
